+++ lib/src/helpers/mixins/serializable_mixin.dart
@@ -1,0 +1,5 @@
+import '../typedefs.dart';
+
+mixin SerializableMixin {
+  JsonMap toJson();
+}